Script now checks if a new schedule was downloaded before running
The SHA 512 of the last used page is compared to the one that was just downloaded. (Overkill? Probably... but... why not?)
The script is also way more configurable, though all configurations are still found hardcoded into the script.

Also, the previous iCal address was invalid. It can be found on:
  https://calendar.google.com/calendar/ical/itca1dvmn55c782volr60339ms%40group.calendar.google.com/public/basic.ics